Removed a couple more of the fake queries (#1856)
